snapd: expect min-entropy-bits to be an integer
Snapd 2.68 returns entropy-bits as a float but min-entropy-bits as an integer.

+++ subiquity/server/snapd/types.py
@@ -351,7 +351,7 @@ class InsufficientEntropyReasons(enum.Enum):
 class InsufficientEntropyDetails:
     reasons: List[InsufficientEntropyReasons]
     entropy_bits: float
-    min_entropy_bits: float
+    min_entropy_bits: int
 
 
 @snapdtype
